+++ runtime/motu/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 edition = "2018"
-name = "motu-runtime"
+name = "snd-firewire-motu-ctl-service"
 version = "0.1.0"
 # For publishing.
 authors = ["Takashi Sakamoto"]
 description = """
-Implementation of service runtime for models based on ASICs of DICE.
+Runtime implementation to control audio and music unit in IEEE 1394 bus which ALSA firewire-motu
+(snd-firewire-motu) supports.
 """
 license = "GPL-3.0-or-later"
 publish = false
